ISOLATE DB LOGIC
