+++ flask_monitoringdashboard/controllers/requests.py
@@ -1,13 +1,14 @@
 import datetime
-
 import numpy
-from sqlalchemy import func, and_
-
 from flask_monitoringdashboard.core.timezone import to_utc_datetime, to_local_datetime
-from flask_monitoringdashboard.database import Request
 from flask_monitoringdashboard.database.count_group import count_requests_per_day, get_value
 from flask_monitoringdashboard.database.endpoint import get_endpoints, get_num_requests
-from flask_monitoringdashboard.database.request import create_time_based_sample_criterion
+from flask_monitoringdashboard.database.request import \
+    get_status_code_frequencies, get_all_request_status_code_counts, get_error_requests_db
+
+
+def get_error_requests(session, endpoint_id, *criterion):
+    return get_error_requests_db(session, endpoint_id, *criterion)
 
 
 def get_num_requests_data(session, start_date, end_date):
@@ -30,22 +31,6 @@ def get_num_requests_data(session, start_date, end_date):
     return {'days': [d.strftime('%Y-%m-%d') for d in days], 'data': data}
 
 
-def get_all_request_status_code_counts(session, endpoint_id):
-    """
-    Gets all the request status code counts.
-
-    :param session: session for the database
-    :param endpoint_id: id for the endpoint
-    :return: A list of tuples in the form of `(status_code, count)`
-    """
-    return (
-        session.query(Request.status_code, func.count(Request.status_code))
-            .filter(Request.endpoint_id == endpoint_id, Request.status_code.isnot(None))
-            .group_by(Request.status_code)
-            .all()
-    )
-
-
 def get_status_code_distribution(session, endpoint_id):
     """
     Gets the distribution of status codes returned by the given endpoint.
@@ -61,44 +46,6 @@ def get_status_code_distribution(session, endpoint_id):
     status_code_counts = get_all_request_status_code_counts(session, endpoint_id)
     total_count = sum(frequency for (_, frequency) in status_code_counts)
     return {status_code: frequency / total_count for (status_code, frequency) in status_code_counts}
-
-
-def get_status_code_frequencies(session, endpoint_id, *criterion):
-    """
-    Gets the frequencies of each status code.
-
-
-    :param session: session for the database
-    :param endpoint_id: id for the endpoint
-    :param criterion: Optional criteria used to file the requests.
-    :return: A dict where the key is the status code and the value is the fraction of requests that returned the status
-    code. Example: a return value of `{ 200: 105, 404: 3 }` means that status code 200 was returned 105 times and
-    404 was returned 3 times.
-    """
-    status_code_counts = session.query(Request.status_code, func.count(Request.status_code)) \
-        .filter(Request.endpoint_id == endpoint_id, Request.status_code.isnot(None), *criterion) \
-        .group_by(Request.status_code).all()
-
-    return dict(status_code_counts)
-
-
-def get_error_requests(session, endpoint_id, *criterion):
-    """
-    Gets all requests that did not return a 200 status code.
-
-    :param session: session for the database
-    :param endpoint_id: ID of the endpoint to be queried
-    :param criterion: Optional criteria used to file the requests.
-    :return:
-    """
-
-    criteria = and_(
-        Request.endpoint_id == endpoint_id,
-        Request.status_code.isnot(None),
-        Request.status_code >= 400,
-        Request.status_code <= 599,
-    )
-    return session.query(Request).filter(criteria, *criterion).all()
 
 
 def get_status_code_frequencies_in_interval(session, endpoint_id, criterion):
