+++ flask_monitoringdashboard/database/auth.py
@@ -2,6 +2,19 @@ from flask_monitoringdashboard.database import User, session_scope
 from flask_monitoringdashboard import config
 
 
+def update_user(new_password, old_password, user_id, user_is_admin):
+    with session_scope() as session:
+        user = session.query(User).filter(User.id == user_id).one()
+        user.is_admin = user_is_admin
+        if old_password:
+            if user.check_password(old_password):
+                user.set_password(new_password)
+                return True
+            else:
+                return False
+        return True
+
+
 def get_user(username, password):
     """Validates the username and password and returns an User-object if both are valid.
     In case the User-table is empty, a user with default credentials is added.
@@ -21,14 +34,28 @@ def get_user(username, password):
     return None
 
 
-def get_all_users(session):
-    users = session.query(User).order_by(User.id).all()
+def add_user(username, is_admin, password):
+    with session_scope() as session:
+        user = User(username=username, is_admin=is_admin)
+        user.set_password(password=password)
+        session.add(user)
+        session.commit()
 
-    return [
-        {
-            'id': user.id,
-            'username': user.username,
-            'is_admin': user.is_admin,
-        }
-        for user in users
-    ]
+
+def delete_user(user_id):
+    with session_scope() as session:
+        session.query(User).filter(User.id == user_id).delete()
+
+
+def get_all_users():
+    with session_scope() as session:
+        users = session.query(User).order_by(User.id).all()
+
+        return [
+            {
+                'id': user.id,
+                'username': user.username,
+                'is_admin': user.is_admin,
+            }
+            for user in users
+        ]
