+++ flask_monitoringdashboard/views/auth.py
@@ -5,8 +5,7 @@ from sqlalchemy.orm.exc import NoResultFound
 
 from flask_monitoringdashboard import blueprint, config
 from flask_monitoringdashboard.core.auth import on_logout, on_login, secure, is_admin, admin_secure
-from flask_monitoringdashboard.database import session_scope, User
-from flask_monitoringdashboard.database.auth import get_user, get_all_users
+from flask_monitoringdashboard.database.auth import get_user, get_all_users, delete_user, add_user, update_user
 
 MAIN_PAGE = config.blueprint_name + '.index'
 BAD_REQUEST_STATUS = 400
@@ -31,10 +30,10 @@ def login():
             on_login(user=user)
             return redirect(url_for(MAIN_PAGE))
     return render_template('fmd_login.html', 
-        blueprint_name=config.blueprint_name, 
-        show_login_banner=config.show_login_banner, 
-        show_login_footer=config.show_login_footer,
-    )
+                           blueprint_name=config.blueprint_name,
+                           show_login_banner=config.show_login_banner,
+                           show_login_footer=config.show_login_footer,
+                           )
 
 
 @blueprint.route('/logout')
@@ -55,8 +54,7 @@ def users_list():
     """
     if not is_admin():
         return jsonify([])
-    with session_scope() as session:
-        return jsonify(get_all_users(session))
+    return jsonify(get_all_users())
 
 
 @blueprint.route('/api/user/delete', methods=['POST'])
@@ -66,8 +64,7 @@ def user_delete():
     user_id = int(request.form['user_id'])
     if flask.session.get(config.link + '_user_id') == user_id:
         return jsonify({'message': 'Cannot delete itself.'}), BAD_REQUEST_STATUS
-    with session_scope() as session:
-        session.query(User).filter(User.id == user_id).delete()
+    delete_user(user_id)
     return 'OK'
 
 
@@ -78,21 +75,17 @@ def user_create():
     username = request.form['username']
     password = request.form['password']
     password2 = request.form['password2']
-    is_admin = request.form['is_admin'] == 'true'
+    user_is_admin = request.form['is_admin'] == 'true'
 
     if password != password2:
         return jsonify({'message': "Passwords don't match."}), BAD_REQUEST_STATUS
 
-    with session_scope() as session:
-        try:
-            user = User(username=username, is_admin=is_admin)
-            user.set_password(password=password)
-            session.add(user)
-            session.commit()
-        except IntegrityError:
-            return jsonify({'message': "Username already exists."}), BAD_REQUEST_STATUS
-        except Exception as e:
-            return jsonify({'message': str(e)}), BAD_REQUEST_STATUS
+    try:
+        add_user(username, user_is_admin, password)
+    except IntegrityError:
+        return jsonify({'message': "Username already exists."}), BAD_REQUEST_STATUS
+    except Exception as e:
+        return jsonify({'message': str(e)}), BAD_REQUEST_STATUS
     return 'OK'
 
 
@@ -101,29 +94,21 @@ def user_create():
 def user_edit():
     """Update the user in the database."""
     user_id = int(request.form['user_id'])
-    is_admin = request.form['is_admin'] == 'true'
-    if flask.session.get(config.link + '_user_id') == user_id and not is_admin:
+    user_is_admin = request.form['is_admin'] == 'true'
+    if flask.session.get(config.link + '_user_id') == user_id and not user_is_admin:
         return jsonify({'message': 'Cannot remove the admin permissions from itself.'}), BAD_REQUEST_STATUS
 
-    with session_scope() as session:
-        try:
-            user = session.query(User).filter(User.id == user_id).one()
-            user.is_admin = is_admin
+    old_password = request.form.get('old_password')
+    new_password = request.form.get('new_password')
 
-            old_password = request.form.get('old_password')
-            if old_password is not None:
-                if user.check_password(old_password):
-                    new_password = request.form['new_password']
-                    new_password2 = request.form['new_password2']
-
-                    if new_password != new_password2:
-                        return jsonify({'message': "Passwords don't match."}), BAD_REQUEST_STATUS
-
-                    user.set_password(new_password)
-                else:
-                    return jsonify({'message': "Old password doesn't match."}), BAD_REQUEST_STATUS
-        except NoResultFound:
-            return jsonify({'message': "User ID doesn't exist."}), BAD_REQUEST_STATUS
-        except Exception as e:
-            return jsonify({'message': str(e)}), BAD_REQUEST_STATUS
+    try:
+        if old_password is not None:
+            if new_password and new_password != request.form['new_password2']:
+                return jsonify({'message': "Passwords don't match."}), BAD_REQUEST_STATUS
+        if not update_user(new_password, old_password, user_id, user_is_admin):
+            return jsonify({'message': "Old password doesn't match."}), BAD_REQUEST_STATUS
+    except NoResultFound:
+        return jsonify({'message': "User ID doesn't exist."}), BAD_REQUEST_STATUS
+    except Exception as e:
+        return jsonify({'message': str(e)}), BAD_REQUEST_STATUS
     return 'OK'
